Wrap do-while loops in repeatable function
The do-while loops were getting a bit rediculous and I can tell I
will need to keep adding more complexity to them. To cut down on
excess code, I wrapped the do-while loops and their conditionals
into a separate function called generateRandomSlot that takes in
the slot of interest and running arrays. It compares to the arrays
within the function and returns a BattlePokemon object ready to be
parsed.