Debug shuffle countdown
The shuffle countdown was not infinite like it was supposed to be
when limited shuffling was not explicitly selected. This was b/c I
forgot to add a line of code to prevent number of rerolls from
decrementing if limitRolls === false. I added a conditional line
to the subtractRoll function and this fixed the bug.